backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
importlib-metadata==6.8.0
Markdown==3.5
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
zipp==3.17.0
